tl;dr:  IT WORKS NOW. Used black magic to fix.
Okay, so it turns out that hidden inputs can only have be of the String type. When I sent the array of arrays of colors to the hidden input, it converted that into a String, which is what was sent to the database as the rows.

This was remedied by adjusting the json_chart_conversion so that it takes this into account.

This also required minor changes to the chart_form to account for the new "chart" to be in this form.

Chart model had all validators and stuff removed for testing purposes.